python_service/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
Flask-Cors==4.0.0
joblib==1.3.2
numpy==1.26.4
playwright==1.43.0
scikit-learn==1.2.2
scipy==1.12.0
threadpoolctl==3.2.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tzdata==2023.4
sentence-transformers==3.1.1
fastapi==0.115.0
uvicorn==0.30.6
python-whois==0.9.4
requests==2.32.3